Removed the ranges from main.js to return to auto-scaling, which is the most flexible option.
Modified the serial connection process in node_client.js to connect to the first device with "Arduino" in the manufacturer name.
That isn't a great way of handling the process, so I'll add a flag for manual selection of port in the future.

Added error catching for the Arduino serialport connection in node_client.js, so errors on reconnect (which are still common) are no longer fatal.

Added install directory with installers for each OS I plan to support.
macOS isn't there yet, but once I write one, I'll add it there.